requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
gunicorn==23.0.0
iniconfig==2.1.0
kombu==5.5.4
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.2
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
